DRTVWR-476, SL-13467: Make LLTrace::BlockTimerStatHandle keys unique.
There are duplicate LLTrace::BlockTimerStatHandle key strings declared in
llsettingsdaycycle.cpp and llsettingswater.cpp -- the only instances of
duplicate BlockTimerStatHandle keys in the viewer code base. SL-13467 tracks
intentional crashes due to duplicate LLInstanceTracker subclass instances with
one of those keys. The simplest experiment to try to eliminate those crashes
is to ensure that every BlockTimerStatHandle in the code base is unique.