FIRE-33106 muted objects are not greyed out in the history when shown
When the show muted text in chat history is enabled the chat from blocked objcets has no differentiation. Adding yet another colour into the mix makes no sense, if we make them both grey then we cannot easily differentiate users/objects so for now I propose simply not persisting this so that when a user accidentally enabels this is goes away again with a relog.