FIRE-33841 - FS Hangs on login
This is triggered by having the stats window open from a previous session. The drawTicks call then manages to get a value_scale = INF call when drawing packet_loss. This causes the main thread to block forever in the loop. the visible symptoms (apart from the hung viewer login) are the event queue filling (reported in the log) because the main thread is no longer consuming.