Fix the bogus, zero cloud scroll rate in default EE sky settings.
Obvioulsy, there has been a typo done when copying WL default sky parameters to EE ones.
This causes "static" and quite unrealistic clouds when this default setting is used as a
base for a new sky setting, and we see this bad static sky resurfacing now with PBR and
its "adjusted" (more like hacked, but this is another story) mid-day sky setting.

Let's fix this typo once and for all in LL's code base (most TPVs have it fixed already,
and this ever since EEP got released).

@LL: please also fix the cloud scroll rate in the PBR mid-day inventory setting accordingly.