[FIRE-35019] - fix memory access error
Fixed crashing issue in LLHUDText::renderText() cause by using mSourceObject without first checking to see if it is not null as the source object could be unloaded when teleporting to different sim.

Was testing going to https://maps.secondlife.com/secondlife/Blue%20Moon%20Bay/188/63/26
for hitching issue and issue was triggered.