[FIRE-35042] - Reverted wording back to Only Coalesced + minor fixes
Based upon feedback, reverted back to using "Only Coalesced" throughout the UI.
Changed the left/right order of the icon and checkbox on the Filter Panel to be better aligned with the other options.
Changed the Only Coalesced check button on the main UI into a universal Show Filters button that shows that any filter is enabled.
Created new Icon image for the filter and added it to the list.
Removed the OnlyCoalesced previous image as it was not needed now.
Cleaned up the code to use the same sync mFilter object to talk between the Finger(Filter) window and the main inventory window.
Added new Show Links filter to the Finder window as a combo box, with three values.
0 - Show Links, 2 - Show Links Only, 1 - Hide Links. This is to match the order on the Gear menu.
Added new feature, where if any filter is turned on, now the bottom inventory show filter button will be highlighted. If you reset or manually turn off the filter it will turn off. This includes setting the Always Show folders check box and the Show Links which did not cause the isNotDefault flag to be true when set.
Works with the second inventory window as well and they don't have conflicts.