FIRE-34749 - editing PBR objects turns them all white
Technically what appears to be happening is that the renderer is identifying the object as wnating to show the BP textures because the BP tab was selected.
This fix tries to correctly select the BP tab by default only when no PBR faces are present. It also triggers a rebuild of the selected objects to force them updated draw because the selection even occurs before the tab change and they would otherwise by displayed using whatever the previous mode was.