Add restore to mBoostLevel (mPrevBoostLevel) for BOOST_SELECTED interactions
Added to the llgltexture a second boost level which will be used to properly restore the boost level after an object is right clicked on as the BOOST_SELECTED code does not restore the value the previous value. This does now and on all the child objects assigned to the object.

Also fixed up the lltextureview.cpp as it had some merge issues from a long time ago and was missed.