FIRE-35198 Limit VRAM texture usage UI control reverts to default value
The manual control for the Limit VRAM texture usage in the Preferences->Graphics->Hardware Settings did not restore the saved value of the slider upload load.

So when the viewer would load, it would indeed load the saved value, but the value was not reflected upon by the slider on the page. This caused issues of when a user went to move the slide back to it's intended position it would set off the bias system as well as make the user not have confidence in that the feature is working.

Added a call to the ser value of the UI control on refresh enable state of the UI floater for the preferences to the saved RenderMaxVRAMBudget.