Fix float typos
There were five files which had float values assigned which were missing the "f" on them so they are in fact a doubles.
These get treated as errors when you change the floating point compile options from Fast to Precise or Strict as the fast path turns them from doubles to floats and suppresses the warnings.