Fix a crash bug and bogus calculations in LLMeshRepoThread::lodReceived() (#4398, #4408)
When trying to update the rigging info for a newly received mesh LOD,
a wrong usage of LLVolume::getNumFaces() is done to get the number of
volume faces, causing the code to iterate over the number of faces in
the underlying LLProfile instead. LLVolume::getNumVolumeFaces() must
be used here.

This fixes a crash bug seen with low LODs in some meshes (when the number
of mesh faces is smaller than the number of faces in the LLProfile), and
also properly updates the rigging info for all mesh faces, as it should,
when the mesh got more faces than the LLProfile.