Address g++ compiler warnings, so we can move to g++14
Avoid uninitialized variable warnings on Linux/g++ by initializing translation_override in LLGLTFLoader.

Force the empty callback in LLPanelRegionTerrainInfo to decay to a function pointer to avoid warnings on Linux.

Remove initialization of mViewerObject in LLReflectionMap.h so that we are standards compliant.

(also paired with commit on fs-build-variables to ignore non-null warnings due to gcc stupidity)