FIRE-36016 - Re-added Store/Restore boost levels of selected objects
Re-adding the feature to store and restore texture boost levels when a user edit's an object.

When you right click on an object. All the textures assigned to it get their boost level set to BOOST_SELECTED.

For LOD textures, once the texture's ProcessStats is completed, it resets the boost level to BOOST_NONE. This can cause the texture to then be subject to Texture Bias when in low memory, where before it was possibly protected as being a higher priority texture.

Now when the boost levels is about to change to BOOST_SELECTED, it gets stored and in ProcessStats, the boost level is restored.

Issue was fixed before but was rolled back. This is just restoring this one fix.