Fix issue where mac was crashing during an attempt to unplug or replug microphone (FS:TJ- Cherry-pick to help fix FIRE-36022 on macOS)
The mac audio device manager was being "helpful" by restarting
playout and recording if the Default device was changed, assuming
the application wouldn't care.
However, we received an update of device change, and attempted to
stop and start playout anyway, causing a conflict.
The fix was simply to not deploy new devices when the device id didn't
change.